fastapi-telemetry/api2/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
prometheus-client
opentelemetry-instrumentation-system-metrics